Refactor tests and optimize units by removing @SpringBootTest annotation
