--- src/test/java/io/bartmilo/student/enrolment/app/ApplicationTests.java
@@ -1,13 +1,0 @@
-package io.bartmilo.student.enrolment.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
